requirements.txt
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
supabase>=0.5.0
python-dotenv>=1.0
httpx>=0.24
rapidfuzz>=2.16

# Dev / testing
pytest>=7.0
